requirements.txt
Flask>=2.0.0
psutil>=5.8.0
python-dotenv>=1.0.0